requirements.txt
absl-py==1.4.0
backcall==0.2.0
cachetools==5.3.1
certifi @ file:///croot/certifi_1671487769961/work/certifi
charset-normalizer==3.1.0
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
Cython==0.29.35
cython-bbox==0.1.3
decorator==5.1.1
faiss==1.7.1
fonttools==4.25.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.21.0
google-auth-oauthlib==0.4.6
grpcio==1.56.0
idna==3.4
importlib-metadata==6.7.0
ipython==7.34.0
jedi==0.18.2
joblib==1.3.1
kiwisolver @ file:///opt/conda/conda-bld/kiwisolver_1638569886207/work
lap==0.4.0
loguru==0.7.0
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib @ file:///croot/matplotlib-suite_1667356714455/work
matplotlib-inline==0.1.6
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626179032232/work
mkl-service==2.4.0
munkres==1.1.4
numpy==1.21.6
nvidia-ml-py==11.525.131
nvitop==1.1.2
oauthlib==3.2.2
opencv-python==4.7.0.72
packaging @ file:///croot/packaging_1671697413597/work
pandas==1.3.5
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
ply==3.11
prettytable==3.7.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.15.1
pyparsing @ file:///opt/conda/conda-bld/pyparsing_1661452539315/work
PyQt5-sip==12.11.0
PySimpleGUI==4.60.5
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2023.3
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.2
sip @ file:///tmp/abs_44cd77b_pu/croots/recipe/sip_1659012365470/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.0
some-package==0.1
tabulate==0.9.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.3.0
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
torch==1.12.0+cu113
torchvision==0.13.0+cu113
tornado @ file:///opt/conda/conda-bld/tornado_1662061693373/work
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.6.3
ultralytics==8.0.124
urllib3==1.26.16
wcwidth==0.2.6
Werkzeug==2.2.3
xlrd==2.0.1
yacs==0.1.8
zipp==3.15.0
